Move DTO classes to model.dto package and update imports accordingly
