+++ backend/src/main/java/com/wmsgroup/neuefische_wms/converter/CategoryOutputDTOConverter.java
@@ -1,6 +1,6 @@
 package com.wmsgroup.neuefische_wms.converter;
 
-import com.wmsgroup.neuefische_wms.dto.CategoryOutputDTO;
+import com.wmsgroup.neuefische_wms.model.dto.CategoryOutputDTO;
 import com.wmsgroup.neuefische_wms.model.Category;
 import lombok.NonNull;
