+++ backend/src/main/java/com/wmsgroup/neuefische_wms/model/dto/CategoryInputDTO.java
@@ -1,4 +1,4 @@
-package com.wmsgroup.neuefische_wms.dto;
+package com.wmsgroup.neuefische_wms.model.dto;
 
 import com.mongodb.lang.Nullable;
 import lombok.NonNull;
