+++ backend/src/main/java/com/wmsgroup/neuefische_wms/model/dto/CategoryOutputDTO.java
@@ -1,4 +1,4 @@
-package com.wmsgroup.neuefische_wms.dto;
+package com.wmsgroup.neuefische_wms.model.dto;
 
 public record CategoryOutputDTO(
         String id,
