+++ backend/src/main/java/com/wmsgroup/neuefische_wms/service/CategoryService.java
@@ -2,8 +2,8 @@ package com.wmsgroup.neuefische_wms.service;
 
 import com.wmsgroup.neuefische_wms.converter.CategoryConverter;
 import com.wmsgroup.neuefische_wms.converter.CategoryOutputDTOConverter;
-import com.wmsgroup.neuefische_wms.dto.CategoryInputDTO;
-import com.wmsgroup.neuefische_wms.dto.CategoryOutputDTO;
+import com.wmsgroup.neuefische_wms.model.dto.CategoryInputDTO;
+import com.wmsgroup.neuefische_wms.model.dto.CategoryOutputDTO;
 import com.wmsgroup.neuefische_wms.repository.CategoryRepository;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
