+++ backend/src/test/java/com/wmsgroup/neuefische_wms/converter/CategoryOutputDTOConverterTest.java
@@ -1,6 +1,6 @@
 package com.wmsgroup.neuefische_wms.converter;
 
-import com.wmsgroup.neuefische_wms.dto.CategoryOutputDTO;
+import com.wmsgroup.neuefische_wms.model.dto.CategoryOutputDTO;
 import com.wmsgroup.neuefische_wms.model.Category;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
